feat: [IDP-724]: Removing dependency from the CI and STO service for execution
