+++ idp-service/src/main/java/io/harness/idp/pipeline/provider/IdpPipelineServiceInfoProvider.java
@@ -12,6 +12,7 @@ import static io.harness.steps.plugin.ContainerStepConstants.PLUGIN;
 
 
 import io.harness.idp.pipeline.step.StepSpecTypeConstants;
+import static io.harness.pms.yaml.YAMLFieldNameConstants.STEP;
 import static io.harness.steps.plugin.ContainerStepConstants.PLUGIN;
 
 import io.harness.beans.steps.StepSpecTypeConstants;
@@ -20,16 +21,24 @@ import io.harness.ci.creator.variables.PluginStepVariableCreator;
 import io.harness.ci.creator.variables.RunStepVariableCreator;
 import io.harness.ci.plan.creator.steps.CIStepsPlanCreator;
 import io.harness.ci.plancreator.GitCloneStepPlanCreator;
+import io.harness.ci.plancreator.InitializeStepPlanCreator;
 import io.harness.ci.plancreator.PluginStepPlanCreator;
 import io.harness.ci.plancreator.RunStepPlanCreator;
+import io.harness.ci.plancreator.V1.BackgroundStepPlanCreatorV1;
+import io.harness.ci.plancreator.V1.TestStepPlanCreator;
+import io.harness.filters.ExecutionPMSFilterJsonCreator;
 import io.harness.idp.pipeline.stages.filtercreator.IDPStageFilterCreator;
 import io.harness.idp.pipeline.stages.filtercreator.IDPStepFilterJsonCreator;
 import io.harness.idp.pipeline.stages.plancreator.IDPStagePlanCreator;
+import io.harness.idp.pipeline.stages.plancreator.IDPStepPlanCreator;
 import io.harness.idp.pipeline.stages.variablecreator.IDPStageVariableCreator;
+import io.harness.idp.pipeline.stages.variablecreator.IDPStepVariableCreator;
+import io.harness.pms.contracts.steps.StepInfo;
 import io.harness.pms.contracts.steps.StepMetaData;
 import io.harness.pms.sdk.core.pipeline.filters.FilterJsonCreator;
 import io.harness.pms.sdk.core.plan.creation.creators.PartialPlanCreator;
 import io.harness.pms.sdk.core.plan.creation.creators.PipelineServiceInfoProvider;
+import io.harness.pms.sdk.core.variables.EmptyVariableCreator;
 import io.harness.pms.sdk.core.variables.VariableCreator;
 import io.harness.pms.utils.InjectorUtils;
 
@@ -37,22 +46,22 @@ import com.google.inject.Inject;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Set;
 
 public class IdpPipelineServiceInfoProvider implements PipelineServiceInfoProvider {
   @Inject InjectorUtils injectorUtils;
+  private static final String LITE_ENGINE_TASK = "liteEngineTask";
 
   @Override
   public List<PartialPlanCreator<?>> getPlanCreators() {
     // Needs to be modified based on steps
     List<PartialPlanCreator<?>> planCreators = new LinkedList<>();
     planCreators.add(new IDPStagePlanCreator());
+    planCreators.add(new IDPStepPlanCreator());
     planCreators.add(new RunStepPlanCreator());
     planCreators.add(new PluginStepPlanCreator());
     planCreators.add(new GitCloneStepPlanCreator());
-    //    planCreators.add(new InitializeStepPlanCreator());
-
-    //    planCreators.add(new RunStepPlanCreatorV1());
-    planCreators.add(new CIStepsPlanCreator());
+    planCreators.add(new InitializeStepPlanCreator());
     injectorUtils.injectMembers(planCreators);
     return planCreators;
   }
@@ -72,9 +81,12 @@ public class IdpPipelineServiceInfoProvider implements PipelineServiceInfoProvid
     // Needs to be modified based on steps
     List<VariableCreator> variableCreators = new ArrayList<>();
     variableCreators.add(new IDPStageVariableCreator());
+    variableCreators.add(new IDPStepVariableCreator());
     variableCreators.add(new RunStepVariableCreator());
     variableCreators.add(new PluginStepVariableCreator());
     variableCreators.add(new GitCloneStepVariableCreator());
+    variableCreators.add(new EmptyVariableCreator(STEP, Set.of(LITE_ENGINE_TASK)));
+
     return variableCreators;
   }
