+++ idp-service/src/main/java/io/harness/idp/pipeline/stages/plancreator/IDPStepPlanCreator.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2021 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
+package io.harness.idp.pipeline.stages.plancreator;
+
+import io.harness.idp.pipeline.stages.utils.IDPCreatorUtils;
+import io.harness.plancreator.steps.GenericStepPMSPlanCreator;
+
+import java.util.Set;
+
+public class IDPStepPlanCreator extends GenericStepPMSPlanCreator {
+  @Override
+  public Set<String> getSupportedStepTypes() {
+    return IDPCreatorUtils.getSupportedSteps();
+  }
+}
